+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed clicking Update Role when no role is selected crashing the app on the single user screen
 - Organizations in the single user screen are now paged
 - Organizations in the single user screen can now be searched
+- Added a health endpoint
 
 ## 0.24.0 - Student Tags
 - Upgraded Rails to 6.0.3.2